### 2025-01-17 #7685 @editorialbot
[PRE REVIEW]: mfpml: Multi-fidelity probabilistic machine learning toolkit
**Submitting author:** @JiaxiangYi96 (<a href="http://orcid.org/0009-0007-9287-4328">Jiaxiang Yi</a>)
**Repository:** https://github.com/JiaxiangYi96/mfpml
**Branch with paper.md** (empty if default branch): pr/2.0
**Version:** pr/2.0
