### 2025-01-21 #7685 @JiaxiangYi96
@arfon Thanks for the kind reminder, I fixed the typo of `peper.bib` by changing it to `paper.bib`
