### 2025-04-05 #7685 @ekourlit
Glad to edit this paper @kyleniemeyer -- I'm quite interested on Bayesian optimization methods.
