### 2025-04-05 #7685 @ekourlit
Greetings @JiaxiangYi96, I will be the editor taking care of your submission. After having an initial look at the repository and the paper, your submission fulfils the minimal requirements to proceed with the review.

Do you already have in mind a couple of individuals that might be willing to review this work for JOSS ? In general, we would appreciate people with domain knowledge in Bayesian optimization. Please just have potential [conflicts of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) in mind. You can consult the automatic suggestions https://github.com/openjournals/joss-reviews/issues/7685#issuecomment-2776103931, or have a look at the [JOSS reviewer database](https://reviewers.joss.theoj.org/).
