### 2025-04-07 #7685 @JiaxiangYi96
Dear @ekourlit,

Thanks for your time on my repo.  I think `zhandawei` and `hxf1228` might be suitable for the topic. In addition, I think the authors of the first two recommended related repos ([BayesO](https://github.com/openjournals/joss-reviews/issues/5320) and [pgGPGO](https://github.com/openjournals/joss-reviews/issues/431))  also have expertise in Bayesian Optimization.
