### 2025-10-25 JiaxiangYi96/mfpml#16 @jungtaekkim
JOSS Review - Manuscript
Hi Authors,

Thank you for contributing to an interesting project on multi-fidelity probabilistic machine learning.

I would like to leave two primary comments on your manuscript:

1. According to the criterion of JOSS submission review, comparisons to existing software should be discussed.  Please add such comparisons.
2. Potential applications should be discussed in the manuscript.  Please describe the potential applications of your software.

This is part of the JOSS submission review (https://github.com/openjournals/joss-reviews/issues/8631).
