### 2025-11-03 2565556
updated manuscript by adding comparison with existed repo and show potential application JOSS Review - Manuscript Fixes #16
