+++ docs/paper/paper.bib
@@ -141,3 +141,30 @@
  pages={2825--2830},
  year={2011}
 }
+
+@misc{gardner2021gpytorchblackboxmatrixmatrixgaussian,
+      title={GPyTorch: Blackbox Matrix-Matrix Gaussian Process Inference with GPU Acceleration}, 
+      author={Jacob R. Gardner and Geoff Pleiss and David Bindel and Kilian Q. Weinberger and Andrew Gordon Wilson},
+      year={2021},
+      eprint={1809.11165},
+      archivePrefix={arXiv},
+      primaryClass={cs.LG},
+      url={https://arxiv.org/abs/1809.11165}, 
+}
+
+@inproceedings{balandat2020botorch,
+  title = {{BoTorch: A Framework for Efficient Monte-Carlo Bayesian Optimization}},
+  author = {Balandat, Maximilian and Karrer, Brian and Jiang, Daniel R. and Daulton, Samuel and Letham, Benjamin and Wilson, Andrew Gordon and Bakshy, Eytan},
+  booktitle = {Advances in Neural Information Processing Systems 33},
+  year = 2020,
+  url = {http://arxiv.org/abs/1910.06403}
+}
+
+@article{saves2024smt,
+         author = {P. Saves and R. Lafage and N. Bartoli and Y. Diouane and J. Bussemaker and T. Lefebvre and J. T. Hwang and J. Morlier and J. R. R. A. Martins},
+         title = {{SMT 2.0: A} Surrogate Modeling Toolbox with a focus on Hierarchical and Mixed Variables Gaussian Processes},
+         journal = {Advances in Engineering Sofware},
+         year = {2024},
+         volume = {188}, 
+         pages = {103571},
+         doi = {https://doi.org/10.1016/j.advengsoft.2023.103571}}
