### 2025-11-23 JiaxiangYi96/mfpml#17 @jungtaekkim
JOSS Review - Documentation
I think the authors can provide more detailed contributing guidelines.

For example, issue templates and more detailed CONTRIBUTING.md and CODE_OF_CONDUCT.md can be added.

This is part of the JOSS submission review (https://github.com/openjournals/joss-reviews/issues/8631).
