### 2025-11-23 JiaxiangYi96/mfpml#18 @jungtaekkim
JOSS Review - Code
I am reviewing your code.

I found that the name of this file seems not correct.

https://github.com/JiaxiangYi96/mfpml/blob/main/src/mfpml/optimization/mf_acqusitions.py

Should it be "mf_acquisitions.py"?

This is part of the JOSS submission review (https://github.com/openjournals/joss-reviews/issues/8631).
