### 2026-01-17 JiaxiangYi96/mfpml#16 @ekourlit
Hi @jungtaekkim ! Are you satisfied with the above update to the manuscript? If so, can you please close this issue?

Best regards,
Vangelis Kourlitis
_Editor of JOSS_
