### 2026-01-18 JiaxiangYi96/mfpml#16 @jungtaekkim
Hi @ekourlit,

Thank you for reminding me of this issue.

Hi the authors,

I think this revision is not enough to show comparisons to existing software including previous literature on multi-fidelity Bayesian optimization and multi-fidelity optimization and potential real-world applications.  Please revise your submission more thoroughly.
